+++ src/editor/hooks/useDeleteNote.ts
@@ -2,6 +2,7 @@ import { useCallback } from 'react';
 import { useCurrentViewContext } from 'context/useCurrentView';
 import deleteBacklinks from 'editor/backlinks/deleteBacklinks';
 import { store, useStore } from 'lib/store';
+import { regDateStr } from 'utils/helper';
 import { writeJsonFile, deleteFile } from 'file/write';
 import { joinPaths } from 'file/util';
 
@@ -37,7 +38,10 @@ export default function useDeleteNote(noteId: string, noteTitle: string) {
     // delete in disk, write to JSON
     const parentDir = store.getState().currentDir;
     if (parentDir) {
-      const toDelPath = await joinPaths(parentDir, [`${noteTitle}.md`]);
+      const isDaily = regDateStr.test(noteTitle);
+      const toDelPath = isDaily 
+        ? await joinPaths(parentDir, ['daily', `${noteTitle}.md`])
+        : await joinPaths(parentDir, [`${noteTitle}.md`]);
       await deleteFile(toDelPath);  // delete file in Disk
       await writeJsonFile(parentDir); // sync the deletion to JSON
     }