+++ src/utils/helper.ts
@@ -2,7 +2,7 @@
 
 // date
 //
-// string: yyyy-mm-dd
+// string: yyyy-mm-dd  or yyyy-m-d
 export const regDateStr = /^([0-9]{4})-(0?[1-9]|1[0-2])-(0?[1-9]|[12][0-9]|3[01])$/;
 
 export function getStrDate(dateStr: string) {